Rename schema files
