+++ src/app/crud/crud_vpcs.py
@@ -5,8 +5,8 @@ from sqlalchemy.orm import load_only, selectinload
 
 from ..models.openlabs_subnet_model import OpenLabsSubnetModel
 from ..models.openlabs_vpc_model import OpenLabsVPCModel
-from ..schemas.openlabs_range_schema import OpenLabsRangeID
-from ..schemas.openlabs_vpc_schema import (
+from ..schemas.template_range_schema import OpenLabsRangeID
+from ..schemas.template_vpc_schema import (
     OpenLabsVPCBaseSchema,
     OpenLabsVPCID,
     OpenLabsVPCSchema,
